add appearance handler
